+++ src/components/ui/index.js
@@ -24,10 +24,10 @@ import MyCardText from "@/components/ui/MyCardText.vue";
 import MySpellText from "@/components/ui/MySpellText.vue";
 import MyWrapper from "@/components/ui/MyWrapper.vue";
 import RaceBody from "@/components/ui/RaceBody.vue";
-import MySkale from "@/components/ui/MySkale.vue";
-import MySizeGrowth from "@/components/ui/MySizeGrowth.vue";
-import MyRange from "@/components/ui/MyRange.vue";
-import MyRangeSize from "@/components/ui/MyRangeSize.vue";
+import MySkale from "@/components/ui/pinia_ui/MySkale.vue";
+import MySizeGrowth from "@/components/ui/pinia_ui/MySizeGrowth.vue";
+import MyRange from "@/components/ui/pinia_ui/MyRange.vue";
+import MyRangeSize from "@/components/ui/pinia_ui/MyRangeSize.vue";
 import MyRangeHight from "@/components/ui/MyRangeHight.vue";
 import MyHeaderCard from "@/components/ui/MyHeaderCard.vue";
 import MyLogoCard from "@/components/ui/MyLogoCard.vue";
